Reimplemented GraphicsContext::drawLineForText().
Old implementation based on drawLine() does not work anymore because change
in the meaning of its arguments in r177686 and following Qt change. Also,
drawLine has complex logic specific to border rendering, and it lacks text-
specific logic present in other ports' implementation. So we should not
use drawLine anymore.

We could rely on drawsLineForText() for implementation like CG and Cairo
do (since it's rendering is fixed now), but its implementation is not
optimal for now. In future we can reconsider this choice.

Change-Id: I9e5d6a939ff62adc807bc269b43c59bb92bfcd31